requirements.txt
fastapi
fastapi-mail
pydantic[email]
sqlalchemy
psycopg2-binary
uvicorn
python-dotenv